renamed shape to modelprimitives
