+++ src/main/scala/game_object_system/graphics_objects/SpriteSheet.scala
@@ -1,0 +1,5 @@
+package game_object_system.graphics_objects
+
+class SpriteSheet {
+
+}
